+++ phpBB2/install/schemas/mysql_schema.sql
@@ -50,7 +50,7 @@ CREATE TABLE phpbb_user_id_sequence (
    singleton tinyint(1) unsigned NOT NULL,
    last_id int(10) unsigned NOT NULL default '0',
    PRIMARY KEY (singleton)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
+) ENGINE=InnoDB ROW_FORMAT=DYNAMIC DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
 
 CREATE TABLE phpbb_user_removals (
    job_id char(32) NOT NULL,
@@ -178,7 +178,7 @@ CREATE TABLE phpbb_config (
     config_name varchar(191) NOT NULL,
     config_value varchar(255) NOT NULL,
     PRIMARY KEY (config_name)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
+) ENGINE=InnoDB ROW_FORMAT=DYNAMIC DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
 
 
 # --------------------------------------------------------
@@ -440,7 +440,7 @@ CREATE TABLE phpbb_sessions (
    PRIMARY KEY (session_id),
    KEY session_user_id (session_user_id),
    KEY session_id_ip_user_id (session_id, session_ip, session_user_id)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
+) ENGINE=InnoDB ROW_FORMAT=DYNAMIC DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
 
 # --------------------------------------------------------
 #